Factor out test for unauthorized requests
